bird_classifier/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.5.0
seaborn>=0.12.0
scikit-learn>=1.0.0
Pillow>=9.0.0
PyYAML>=6.0
dvc>=3.0.0
numpy>=1.21.0
faiss-cpu>=1.7.4
gradio>=4.0.0
tqdm>=4.65.0
